Add files via upload
Welcome to Cibo MVC Framework
Cibo (pronounced Shi-bo) is a minimal framework for building web apps named after Cibo from the blame manga series.

CIBO comes ready with:
- login systen
Coming Soon:
- breadcrumb
- CRUD
- search
- template
Who is this framework for?
Anyone who is new to the MVC pattern and wants to learn building web applications.

To get started download the framework, and start a local server. There are instructions on the homepage.